refactor: split cart client component
